mcp-servers/requirements.txt
# MCP Framework
mcp>=1.0.0
pydantic>=2.0.0

# Weather Server
requests>=2.26.0
python-dotenv>=0.19.0

# Google Drive Server  
google-api-python-client>=2.31.0
google-auth>=2.6.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0

# Development/Testing
pytest>=6.2.5
pytest-asyncio>=0.21.0